+++ .github/workflows/ci.yaml
@@ -41,7 +41,7 @@ jobs:
       - name: Run Unit Tests
         shell: bash -l {0}
         run: |
-          coverage run --rcfile=coverage.toml -m pytest --verbose esnpy/test/driver.py esnpy/test/matrix.py esnpy/test/esn.py esnpy/test/lazy.py esnpy/test/xdata.py
+          coverage run --rcfile=coverage.toml -m pytest --verbose xesn/test/driver.py xesn/test/matrix.py xesn/test/esn.py xesn/test/lazy.py xesn/test/xdata.py
       - name: Show coverage report
         shell: bash -l {0}
         run: |
