+++ coverage.toml
@@ -1,5 +1,5 @@
 [tool.coverage.run]
 omit = [
-    "esnpy/test/*.py",
+    "xesn/test/*.py",
     "**/__init__.py",
 ]
