+++ docs/.gitignore
@@ -1,1 +1,1 @@
-source/esnpy
+source/xesn
