+++ docs/environment.yaml
@@ -1,4 +1,4 @@
-name: esn_docs 
+name: xesn_docs 
 channels:
     - conda-forge
 dependencies:
