+++ docs/source/index.rst
@@ -1,10 +1,10 @@
-.. esnpy documentation master file, created by
+.. xesn documentation master file, created by
    sphinx-quickstart on Fri Oct  6 14:42:27 2023.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to esnpy's documentation!
-=================================
+xesn Documentation
+==================
 
 .. toctree::
    :maxdepth: 1
