+++ docs/source/notebooks/xesn.mplstyle
