+++ xesn/test/driver.py
@@ -11,9 +11,9 @@ from shutil import rmtree
 
 import dask.array as darray
 
-from esnpy.driver import Driver
-from esnpy.test.xdata import test_data
-from esnpy.test.esn import test_data as eager_data
+from xesn.driver import Driver
+from xesn.test.xdata import test_data
+from xesn.test.esn import test_data as eager_data
 
 @pytest.fixture(scope="function")
 def config_dict():
